Re-implement optimizations from codex branch with proper fixes
- dirctl.c: Add buffer overflow protection (NameLength > MAX_PATH check)
- ea.c: Add EA size validation (name_len <= 255, data_size <= 65535)
- volinfo.c: Fix integer overflow (cast ext3_blocks_count to ULONGLONG)
- flush.c: Use shared lock instead of exclusive for Ext2FlushFiles (avoids deadlock)
- ext4_csum.c: Add ext4_chksum_batch() utility for sequential batch CRC
- htree.c: Re-implement htree directory cache with proper KSPIN_LOCK threading
  protection (KeAcquireInStackQueuedSpinLock) and thread-safe init guard
- init.c: Call htree_cache_init() during driver startup

+++ Ext4Fsd/volinfo.c
@@ -130,9 +130,9 @@ Ext2QueryVolumeInformation (IN PEXT2_IRP_CONTEXT IrpContext)
 
             FsSizeInfo = (PFILE_FS_SIZE_INFORMATION) Buffer;
             FsSizeInfo->TotalAllocationUnits.QuadPart =
-                ext3_blocks_count(SUPER_BLOCK);
+                (ULONGLONG)ext3_blocks_count(SUPER_BLOCK);
             FsSizeInfo->AvailableAllocationUnits.QuadPart =
-                ext3_free_blocks_count(SUPER_BLOCK);
+                (ULONGLONG)ext3_free_blocks_count(SUPER_BLOCK);
             FsSizeInfo->SectorsPerAllocationUnit =
                 Vcb->BlockSize / Vcb->DiskGeometry.BytesPerSector;
             FsSizeInfo->BytesPerSector =
